requirements.txt
# Test dependencies for functional tests
pytest
Pillow
pytesseract
websocket-client
